# docs/openapi-leanschool.yaml
openapi: 3.0.3
info:
  title: leanschool API
  description: CRUD data layer for school management (receipts, locations, persons,
    classes, etc.).
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: local
paths:
  /addresses:
    get:
      summary: List addresses
      operationId: listAddresss
      responses:
        '200':
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a addresse
      operationId: createAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '201':
          description: Created Address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /buildings:
    get:
      summary: List buildings
      operationId: listBuildings
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Building'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a building
      operationId: createBuilding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildingRequest'
      responses:
        '201':
          description: Created Building
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /cities:
    get:
      summary: List cities
      operationId: listCitys
      responses:
        '200':
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a citie
      operationId: createCity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
      responses:
        '201':
          description: Created City
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /curricula:
    get:
      summary: List curricula
      operationId: listCurriculums
      responses:
        '200':
          description: List of curricula
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Curriculum'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a curricul
      operationId: createCurriculum
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Curriculum'
      responses:
        '201':
          description: Created Curriculum
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curriculum'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /exams:
    get:
      summary: List exams
      operationId: listExams
      responses:
        '200':
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exam'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a exam
      operationId: createExam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamRequest'
      responses:
        '201':
          description: Created Exam
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exam'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /grades:
    get:
      summary: List grades
      operationId: listGrades
      responses:
        '200':
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grade'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a grade
      operationId: createGrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeRequest'
      responses:
        '201':
          description: Created Grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /guardians:
    get:
      summary: List guardians
      operationId: listGuardians
      responses:
        '200':
          description: List of guardians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Guardian'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a guardian
      operationId: createGuardian
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardianRequest'
      responses:
        '201':
          description: Created Guardian
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardian'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /lessons:
    get:
      summary: List lessons
      operationId: listLessons
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a lesson
      operationId: createLesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonRequest'
      responses:
        '201':
          description: Created Lesson
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /locations:
    get:
      summary: List locations
      operationId: listLocations
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a location
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '201':
          description: Created Location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /persons:
    get:
      summary: List persons
      operationId: listPersons
      responses:
        '200':
          description: List of persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a person
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonRequest'
      responses:
        '201':
          description: Created Person
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /postal-codes:
    get:
      summary: List postal codes
      operationId: listPostalCodes
      responses:
        '200':
          description: List of postal codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostalCode'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a postal code
      operationId: createPostalCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostalCode'
      responses:
        '201':
          description: Created PostalCode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalCode'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /rooms:
    get:
      summary: List rooms
      operationId: listRooms
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a room
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomRequest'
      responses:
        '201':
          description: Created Room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /school-classes:
    get:
      summary: List school classes
      operationId: listSchoolClasss
      responses:
        '200':
          description: List of school classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolClass'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a school classe
      operationId: createSchoolClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolClassRequest'
      responses:
        '201':
          description: Created SchoolClass
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolClass'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /school-years:
    get:
      summary: List school years
      operationId: listSchoolYears
      responses:
        '200':
          description: List of school years
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolYear'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a school year
      operationId: createSchoolYear
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolYear'
      responses:
        '201':
          description: Created SchoolYear
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolYear'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /students:
    get:
      summary: List students
      operationId: listStudents
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a student
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRequest'
      responses:
        '201':
          description: Created Student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /subjects:
    get:
      summary: List subjects
      operationId: listSubjects
      responses:
        '200':
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a subject
      operationId: createSubject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRequest'
      responses:
        '201':
          description: Created Subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /teachers:
    get:
      summary: List teachers
      operationId: listTeachers
      responses:
        '200':
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a teacher
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherRequest'
      responses:
        '201':
          description: Created Teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  schemas:
    Address:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: string
          example: 1a
        postalCode:
          $ref: '#/components/schemas/PostalCode'
        street:
          type: string
          example: Bahnhofstrasse
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    AddressRequest:
      type: object
      properties:
        number:
          type: string
          example: 1a
        postalCode:
          $ref: '#/components/schemas/PostalCodeRef'
        street:
          type: string
          example: Bahnhofstrasse
        version:
          type: integer
          example: 0
      required:
      - version
    Building:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        location:
          $ref: '#/components/schemas/Location'
        name:
          type: string
          example: Main Building
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - version
    BuildingRequest:
      type: object
      properties:
        location:
          $ref: '#/components/schemas/EntityRef'
        name:
          type: string
          example: Main Building
        version:
          type: integer
          example: 0
      required:
      - name
      - version
    City:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Zurich
        postalCodes:
          type: array
          items:
            $ref: '#/components/schemas/PostalCode'
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - postalCodes
      - version
    CityRequest:
      type: object
      properties:
        name:
          type: string
          example: Zurich
        postalCodes:
          type: array
          items:
            $ref: '#/components/schemas/PostalCodeRef'
        version:
          type: integer
          example: 0
      required:
      - name
      - postalCodes
      - version
    Curriculum:
      type: object
      properties:
        activeFrom:
          type: string
          format: date-time
        activeSince:
          type: string
          format: date-time
        activeUntil:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: National Curriculum 2024
        version:
          type: integer
          example: 0
      required:
      - id
      - version
      - name
...
    Exam:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        schoolClass:
          $ref: '#/components/schemas/SchoolClass'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    ExamRequest:
      type: object
      properties:
        schoolClass:
          $ref: '#/components/schemas/EntityRef'
        version:
          type: integer
          example: 0
      required:
      - version
    Grade:
      type: object
      properties:
        exam:
          $ref: '#/components/schemas/Exam'
        grade:
          type: number
          format: double
          example: 5.5
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        student:
          $ref: '#/components/schemas/Student'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    GradeRequest:
      type: object
      properties:
        exam:
          $ref: '#/components/schemas/EntityRef'
        grade:
          type: number
          format: double
          example: 5.5
        student:
          $ref: '#/components/schemas/EntityRef'
        version:
          type: integer
          example: 0
      required:
      - version
    Guardian:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - version
    GuardianRequest:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/EntityRef'
        dateOfBirth:
          type: string
          format: date-time
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - name
      - prename
      - version
    Lesson:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        schoolClass:
          $ref: '#/components/schemas/SchoolClass'
        subject:
          $ref: '#/components/schemas/Subject'
        teacher:
          $ref: '#/components/schemas/Teacher'
        version:
          type: integer
          example: 0
        room:
          $ref: '#/components/schemas/Room'
        dayOfWeek:
          type: integer
          example: 1
          description: 1=Monday, 5=Friday
        period:
          type: integer
          example: 1
        startTime:
          type: string
          example: 08:00
        endTime:
          type: string
          example: 08:45
      required:
      - id
      - version
    LessonRequest:
      type: object
      properties:
        schoolClass:
          $ref: '#/components/schemas/EntityRef'
        subject:
          $ref: '#/components/schemas/EntityRef'
        teacher:
          $ref: '#/components/schemas/EntityRef'
        version:
          type: integer
          example: 0
        room:
          $ref: '#/components/schemas/EntityRef'
        dayOfWeek:
          type: integer
        period:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
      required:
      - version
    Location:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        lat:
          type: number
          format: double
          example: 47.3769
        lon:
          type: number
          format: double
          example: 8.5417
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    Person:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - version
    PersonRequest:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/EntityRef'
        dateOfBirth:
          type: string
          format: date-time
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - name
      - prename
      - version
    PostalCode:
      type: object
      properties:
        city:
          type: string
          example: Zurich
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: integer
          example: 8001
        version:
          type: integer
          example: 0
      required:
      - id
      - city
      - version
      - number
...
    Room:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/Building'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Room 101
        version:
          type: integer
          example: 0
        roomType:
          $ref: '#/components/schemas/RoomType'
      required:
      - id
      - name
      - version
    RoomRequest:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/EntityRef'
        name:
          type: string
          example: Room 101
        version:
          type: integer
          example: 0
      required:
      - name
      - building
      - version
...
    SchoolYear:
      type: object
      properties:
        from:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        to:
          type: string
          format: date-time
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
  securitySchemes:
    keycloak:
      type: openIdConnect
      description: Keycloak OpenID Connect — JWT bearer token with realm_access roles
        (leanschool_read, leanschool_write)
      openIdConnectUrl: http://localhost:8180/realms/leanschool/.well-known/openid-configuration
security:
- keycloak: []